+++ core/model/cachedbuildings.go
@@ -17,7 +17,7 @@ package model
 import (
 	"time"
 
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
 const (